server/requirements.txt
alembic==1.5.5
aniso8601==9.0.0
astroid==2.5
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
colorama==0.4.4
dnspython==2.1.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.4.4
Flask-JWT-Extended==4.0.2
Flask-RESTful==0.3.8
Flask-WTF==0.14.3
idna==3.1
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.5.2
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==2.0.3
pycparser==2.20
python-dotenv==0.15.0
python-editor==1.0.4
PyJWT==2.0.1
pylint==2.7.0
pytz==2021.1
six==1.15.0
SQLAlchemy==1.3.23
toml==0.10.2
typed-ast==1.4.2
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3